fix conflicts with change of integer registration to Long registration in User class - tests need revision

+++ src/main/java/br/edu/ifpb/dac/ssp/controller/SchedulingController.java
@@ -164,7 +164,7 @@ public class SchedulingController {
 	}
 	
 	@PatchMapping("/participation/add/{id}")
-	public ResponseEntity addParticipant(@PathVariable Integer id, @RequestBody Integer userRegistration) {
+	public ResponseEntity addParticipant(@PathVariable Integer id, @RequestBody Long userRegistration) {
 		try {
 			User user = userService.findByRegistration(userRegistration).orElse(null);
 
@@ -179,7 +179,7 @@ public class SchedulingController {
 	}
 	
 	@PatchMapping("/participation/remove/{id}")
-	public ResponseEntity removeParticipant(@PathVariable Integer id, @RequestBody Integer userRegistration) {
+	public ResponseEntity removeParticipant(@PathVariable Integer id, @RequestBody Long userRegistration) {
 		try {
 			User user = userService.findByRegistration(userRegistration).orElse(null);
